+++ src/Logger.h
@@ -30,7 +30,7 @@ public:
 	void Start();
 	void Stop();
 	bool IsEnabled();
-	int GetRate();
+	float GetRate();
 
 	void AddtoBuffer(string name, string value);//write data to a buffer to be written to a file later.
 	void AddtoBuffer(string name, double value);
